+++ NEWS.md
@@ -18,7 +18,8 @@ eco-system and let to define it globally with `set_flextable_defaults()`.
 providing a *scroll* value for argument `opts_html`: 
 `opts_html = list(scroll = list(height = "500px",freeze_first_column = TRUE))`.
 - new function `grid_chunk()` to let users add grid graphics
-- functions `add_header_lines()` and `add_footer_lines()` 
+- functions `add_header_row()`, `add_footer_row()`, `add_body_row()`,
+`add_header_lines()` and `add_footer_lines()` 
 now supports formatted paragraph made with `as_paragraph()`.
 - captions: support for simple text in addition to `as_paragraph()`